+++ panther_gazebo/include/panther_gazebo/gui/e_stop.hpp
