+++ panther_gazebo/src/gui/EStop.qrc
